Partial changes in CRUD API Posts and Recipes

+++ src/main/java/com/sinter/cookspire/dto/RepliesDTO.java
@@ -18,10 +18,10 @@ public class RepliesDTO {
     private long commentId;
 
     @Nullable
-    private long likes;
+    private int likes;
 
     @Nullable
-    private long dislikes;
+    private int dislikes;
 
     @Nullable
     private LocalDateTime createdOn;
